requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
joblib>=1.3.0
tqdm>=4.64.0
pyyaml>=6.0
flask>=2.3.0
gunicorn>=21.2.0
rich>=13.0.0
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
